refactor: 뉴스 수집 컴포넌트 네이밍 변경 및 패키지 이동
